Patch dwm with status2d and systray embedded patch
status2d
--------

Status2d allows colors and rectangle drawing in your DWM status bar. See
below an example of my status bar with multi-cpu and battery.

systray
-------

A simple system tray implementation. Multi-monitor is also supported.
The tray follows the selected monitor. The system tray is a segment of a
desktop where some applications can leave an icon with the app’s
functions to quickly interact with a program that is running in the
background.